+++ src/framework/containers/particles.h
@@ -98,7 +98,7 @@ namespace ntt {
      * @param cooling The cooling mechanism assigned for the species
      * @param npld The number of payloads for the species
      */
-    Particles(int                index,
+    Particles(unsigned short     index,
               const std::string& label,
               float              m,
               float              ch,
